bugfix: properly check tracked files for uncommitted changes
Commit d061319 added the good idea of skipping over untracked files
when checking for uncommited changes, but introduced a regression
where functionally NO files are checked for uncommited changes.

The previous command was equivalent to the following:
`git status -s -u no`
`git status -s --untracked-files=all -- no`

Here "no" is treated as a filename, so only "./no" is checked, and
any file changes that would normally be tracked are ignored.

ref: https://git-scm.com/docs/git-status#Documentation/git-status.txt--ultmodegt